+++ extensions/bundles/vcpe/src/main/java/org/opennaas/extensions/vcpe/manager/templates/TemplateSelector.java
@@ -4,7 +4,7 @@
 package org.opennaas.extensions.vcpe.manager.templates;
 
 import org.opennaas.extensions.vcpe.manager.VCPENetworkManagerException;
-import org.opennaas.extensions.vcpe.manager.templates.sp.MultipleProviderTemplate;
+import org.opennaas.extensions.vcpe.manager.templates.mp.MultipleProviderTemplate;
 import org.opennaas.extensions.vcpe.manager.templates.sp.SingleProviderTemplate;
 
 /**
